Integrate ConnectionAvailabilityAwareActivity into main activity.
Makes it easier to refactor fragment creation into new class.
